moved all classes in separate files in base_classes folder
