+++ config/config_base.yaml
@@ -14,7 +14,7 @@ proc:
   # It can be either 'all' for processing all the epochs,
   # A list of two elements for processing a range of epoches
   # or a list with all the epoche numbers to be processed
-  epoch_to_process: "all" # [0, 3] # # [0] #
+  epoch_to_process: [0, 2] #"all" # [0, 3] # # [0] #
 
   #- Switches to find and track matches
   do_matching: true #true #
