+++ lib/tiles.py
@@ -30,9 +30,7 @@ import numpy as np
 from pathlib import Path
 from typing import List, Union
 
-# from lib.classes import Imageds
-from classes import Imageds
-from base_classes.images import Image
+from base_classes.images import Image, Imageds
 
 
 def generateTiles(
