+++ lib/utils/collimate_points.m
