+++ lib/utils/collimate_points_in_dir.m
